backend/requirements.txt
django
djangorestframework
mysqlclient
pandas
numpy
scikit-learn
django-cors-headers
requests